+++ src/pages/home/report/ReportActionItem.js
@@ -2,8 +2,6 @@ import _ from 'underscore';
 import React, {Component} from 'react';
 import {View} from 'react-native';
 import PropTypes from 'prop-types';
-import {withOnyx} from 'react-native-onyx';
-import ONYXKEYS from '../../../ONYXKEYS';
 import ReportActionPropTypes from './ReportActionPropTypes';
 import {
     getReportActionItemStyle,
@@ -29,16 +27,6 @@ const propTypes = {
 
     // Should we display the new indicator on top of the comment?
     shouldDisplayNewIndicator: PropTypes.bool.isRequired,
-
-    // Information about the network
-    network: PropTypes.shape({
-        // Is the network currently offline or not
-        isOffline: PropTypes.bool,
-    }),
-};
-
-const defaultProps = {
-    network: {isOffline: false},
 };
 
 class ReportActionItem extends Component {
@@ -107,18 +95,8 @@ class ReportActionItem extends Component {
                             )}
                             <View style={getReportActionItemStyle(hovered)}>
                                 {!this.props.displayAsGroup
-                                    ? (
-                                        <ReportActionItemSingle
-                                            action={this.props.action}
-                                            isOffline={this.props.network.isOffline}
-                                        />
-                                    )
-                                    : (
-                                        <ReportActionItemGrouped
-                                            action={this.props.action}
-                                            isOffline={this.props.network.isOffline}
-                                        />
-                                    )}
+                                    ? <ReportActionItemSingle action={this.props.action} />
+                                    : <ReportActionItemGrouped action={this.props.action} />}
                             </View>
                             <View style={getMiniReportActionContextMenuWrapperStyle(this.props.displayAsGroup)}>
                                 <ReportActionContextMenu
@@ -160,10 +138,5 @@ class ReportActionItem extends Component {
 }
 
 ReportActionItem.propTypes = propTypes;
-ReportActionItem.defaultProps = defaultProps;
 
-export default withOnyx({
-    network: {
-        key: ONYXKEYS.NETWORK,
-    },
-})(ReportActionItem);
+export default ReportActionItem;
