+++ src/pages/home/report/ReportActionItemMessage.js
@@ -2,6 +2,8 @@ import React from 'react';
 import {View} from 'react-native';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
+import {withOnyx} from 'react-native-onyx';
+import ONYXKEYS from '../../../ONYXKEYS';
 import styles from '../../../styles/styles';
 import ReportActionItemFragment from './ReportActionItemFragment';
 import ReportActionPropTypes from './ReportActionPropTypes';
@@ -10,13 +12,20 @@ const propTypes = {
     // The report action
     action: PropTypes.shape(ReportActionPropTypes).isRequired,
 
-    // Is the network currently offline or not
-    isOffline: PropTypes.bool.isRequired,
+    // Information about the network
+    network: PropTypes.shape({
+        // Is the network currently offline or not
+        isOffline: PropTypes.bool,
+    }),
 };
 
-const ReportActionItemMessage = ({action, isOffline}) => {
+const defaultProps = {
+    network: {isOffline: false},
+};
+
+const ReportActionItemMessage = ({action, network}) => {
     // reportActionID is only present when the action is saved onto server.
-    const isUnsent = isOffline && action.loading && !action.reportActionID;
+    const isUnsent = network.isOffline && action.loading && !action.reportActionID;
     return (
         <View style={[styles.chatItemMessage, isUnsent && styles.chatItemUnsentMessage]}>
             {_.map(_.compact(action.message), (fragment, index) => (
@@ -32,6 +41,11 @@ const ReportActionItemMessage = ({action, isOffline}) => {
 };
 
 ReportActionItemMessage.propTypes = propTypes;
+ReportActionItemMessage.defaultProps = defaultProps;
 ReportActionItemMessage.displayName = 'ReportActionItemMessage';
 
-export default ReportActionItemMessage;
+export default withOnyx({
+    network: {
+        key: ONYXKEYS.NETWORK,
+    },
+})(ReportActionItemMessage);
